Upgrade to 1.3-beta16 from upstream

+++ src/app/extensions.cpp
@@ -1,5 +1,5 @@
 // Aseprite
-// Copyright (C) 2020  Igara Studio S.A.
+// Copyright (C) 2020-2022  Igara Studio S.A.
 // Copyright (C) 2017-2018  David Capello
 //
 // This program is distributed under the terms of
@@ -257,15 +257,21 @@ void Extension::executeExitActions()
 #endif // ENABLE_SCRIPTING
 }
 
+void Extension::addKeys(const std::string& id, const std::string& path)
+{
+  m_keys[id] = path;
+  updateCategory(Category::Keys);
+}
+
 void Extension::addLanguage(const std::string& id, const std::string& path)
 {
   m_languages[id] = path;
   updateCategory(Category::Languages);
 }
 
-void Extension::addTheme(const std::string& id, const std::string& path)
+void Extension::addTheme(const std::string& id, const std::string& path, const std::string& variant)
 {
-  m_themes[id] = path;
+  m_themes[id] = ThemeInfo(path, variant);
   updateCategory(Category::Themes);
 }
 
@@ -347,7 +353,7 @@ void Extension::enable(const bool state)
 #endif // ENABLE_SCRIPTING
 }
 
-void Extension::uninstall()
+void Extension::uninstall(const DeletePluginPref delPref)
 {
   if (!m_isInstalled)
     return;
@@ -360,14 +366,15 @@ void Extension::uninstall()
         m_name.c_str(), m_path.c_str());
 
   // Remove all files inside the extension path
-  uninstallFiles(m_path);
-  ASSERT(!base::is_directory(m_path));
+  uninstallFiles(m_path, delPref);
+  ASSERT(!base::is_directory(m_path) || delPref == DeletePluginPref::kNo);
 
   m_isEnabled = false;
   m_isInstalled = false;
 }
 
-void Extension::uninstallFiles(const std::string& path)
+void Extension::uninstallFiles(const std::string& path,
+                               const DeletePluginPref delPref)
 {
 #if 1 // Read the list of files to be uninstalled from __info.json file
 
@@ -392,11 +399,17 @@ void Extension::uninstallFiles(const std::string& path)
     }
   }
 
-  // Delete __pref.lua file
+  // Delete __pref.lua file (only if specified, e.g. if the user is
+  // updating the extension, the preferences should be kept).
+  bool hasPrefFile = false;
   {
     std::string fn = base::join_path(path, kPrefLua);
-    if (base::is_file(fn))
-      base::delete_file(fn);
+    if (base::is_file(fn)) {
+      if (delPref == DeletePluginPref::kYes)
+        base::delete_file(fn);
+      else
+        hasPrefFile = true;
+    }
   }
 
   std::sort(installedDirs.begin(),
@@ -421,7 +434,8 @@ void Extension::uninstallFiles(const std::string& path)
   }
 
   TRACE("EXT: Deleting extension directory '%s'\n", path.c_str());
-  base::remove_directory(path);
+  if (!hasPrefFile)
+    base::remove_directory(path);
 
 #else // The following code delete the whole "path",
       // we prefer the __info.json approach.
@@ -433,7 +447,7 @@ void Extension::uninstallFiles(const std::string& path)
       base::delete_file(fn);
     }
     else if (base::is_directory(fn)) {
-      uninstallFiles(fn);
+      uninstallFiles(fn, deleteUserPref);
     }
   }
 
@@ -456,8 +470,10 @@ bool Extension::isDefaultTheme() const
 
 void Extension::updateCategory(const Category newCategory)
 {
-  if (m_category == Category::None)
+  if (m_category == Category::None ||
+      m_category == Category::Keys) {
     m_category = newCategory;
+  }
   else if (m_category != newCategory)
     m_category = Category::Multiple;
 }
@@ -789,7 +805,7 @@ std::string Extensions::themePath(const std::string& themeId)
 
     auto it = ext->themes().find(themeId);
     if (it != ext->themes().end())
-      return it->second;
+      return it->second.path;
   }
   return std::string();
 }
@@ -852,9 +868,10 @@ void Extensions::enableExtension(Extension* extension, const bool state)
   generateExtensionSignals(extension);
 }
 
-void Extensions::uninstallExtension(Extension* extension)
+void Extensions::uninstallExtension(Extension* extension,
+                                    const DeletePluginPref delPref)
 {
-  extension->uninstall();
+  extension->uninstall(delPref);
   generateExtensionSignals(extension);
 
   auto it = std::find(m_extensions.begin(),
@@ -1016,6 +1033,24 @@ Extension* Extensions::loadExtension(const std::string& path,
 
   auto contributes = json["contributes"];
   if (contributes.is_object()) {
+    // Keys
+    auto keys = contributes["keys"];
+    if (keys.is_array()) {
+      for (const auto& key : keys.array_items()) {
+        std::string keyId = key["id"].string_value();
+        std::string keyPath = key["path"].string_value();
+
+        // The path must be always relative to the extension
+        keyPath = base::join_path(path, keyPath);
+
+        LOG("EXT: New keyboard shortcuts '%s' in '%s'\n",
+            keyId.c_str(),
+            keyPath.c_str());
+
+        extension->addKeys(keyId, keyPath);
+      }
+    }
+
     // Languages
     auto languages = contributes["languages"];
     if (languages.is_array()) {
@@ -1026,7 +1061,7 @@ Extension* Extensions::loadExtension(const std::string& path,
         // The path must be always relative to the extension
         langPath = base::join_path(path, langPath);
 
-        LOG("EXT: New language '%s' in '%s'\n",
+        LOG("EXT: New language id=%s path=%s\n",
             langId.c_str(),
             langPath.c_str());
 
@@ -1040,15 +1075,17 @@ Extension* Extensions::loadExtension(const std::string& path,
       for (const auto& theme : themes.array_items()) {
         std::string themeId = theme["id"].string_value();
         std::string themePath = theme["path"].string_value();
+        std::string themeVariant = theme["variant"].string_value();
 
         // The path must be always relative to the extension
         themePath = base::join_path(path, themePath);
 
-        LOG("EXT: New theme '%s' in '%s'\n",
+        LOG("EXT: New theme id=%s path=%s variant=%s\n",
             themeId.c_str(),
-            themePath.c_str());
+            themePath.c_str(),
+            themeVariant.c_str());
 
-        extension->addTheme(themeId, themePath);
+        extension->addTheme(themeId, themePath, themeVariant);
       }
     }
 
@@ -1062,7 +1099,7 @@ Extension* Extensions::loadExtension(const std::string& path,
         // The path must be always relative to the extension
         palPath = base::join_path(path, palPath);
 
-        LOG("EXT: New palette '%s' in '%s'\n",
+        LOG("EXT: New palette id=%s path=%s\n",
             palId.c_str(),
             palPath.c_str());
 
@@ -1083,7 +1120,7 @@ Extension* Extensions::loadExtension(const std::string& path,
         // The path must be always relative to the extension
         matPath = base::join_path(path, matPath);
 
-        LOG("EXT: New dithering matrix '%s' in '%s'\n",
+        LOG("EXT: New dithering matrix id=%s path=%s\n",
             matId.c_str(),
             matPath.c_str());
 
@@ -1103,7 +1140,7 @@ Extension* Extensions::loadExtension(const std::string& path,
         // The path must be always relative to the extension
         scriptPath = base::join_path(path, scriptPath);
 
-        LOG("EXT: New script '%s'\n", scriptPath.c_str());
+        LOG("EXT: New script path=%s\n", scriptPath.c_str());
 
         extension->addScript(scriptPath);
       }
@@ -1116,7 +1153,7 @@ Extension* Extensions::loadExtension(const std::string& path,
       // The path must be always relative to the extension
       scriptPath = base::join_path(path, scriptPath);
 
-      LOG("EXT: New script '%s'\n", scriptPath.c_str());
+      LOG("EXT: New script path=%s\n", scriptPath.c_str());
 
       extension->addScript(scriptPath);
     }
@@ -1130,6 +1167,7 @@ Extension* Extensions::loadExtension(const std::string& path,
 
 void Extensions::generateExtensionSignals(Extension* extension)
 {
+  if (extension->hasKeys()) KeysChange(extension);
   if (extension->hasLanguages()) LanguagesChange(extension);
   if (extension->hasThemes()) ThemesChange(extension);
   if (extension->hasPalettes()) PalettesChange(extension);
